썸네일을 드래그앤 드롭해서 바꿀 수 있도록 변경함.
