+++ assets/template/detail.html
@@ -8,7 +8,6 @@
 <script src="/assets/js/jquery-3.1.1.min.js"></script>
 <script src="/assets/bootstrap-4/js/bootstrap.bundle.min.js"></script>
 <script src="/assets/js/input.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
 <script>
     // Tooltip을 띄운다.
     $(function () {
