--- assets/template/editItem.html
@@ -1,28 +1,0 @@
-{{define "edititem" }}
-{{template "headBootstrap"}}
-{{template "navbar" .}}
-
-<div class="col-lg-6 col-md-8 col-sm-12 mx-auto">		
-	<form action="/edititem-submit" method="POST" enctype="multipart/form-data">
-		<input type="hidden" name="project" value={{.Project.ID}}>
-		<input type="hidden" name="id" value={{.Item.ID}}>
-		<div class="form-group pt-5">
-			<label>{{.Project.ID}} 프로젝트 {{.Item.ID}} 아이템의 Thumbnail을 변경합니다</label>
-			<input type="file" name="Thumbnail" class="form-control">
-			<small class="form-text text-mute">핸드폰을 이용해서 사진촬영으로 썸네일을 업로드 할 경우 가로모드로 촬영해주세요.</small>
-		</div>
-		<div class="row">
-			<div class="mx-auto text-center">
-				<button class="btn btn-outline-warning my-3" id="button">Edit</button>
-			</div>
-		</div>
-	</form>
-</div>
-
-{{template "footerBootstrap" .}}
-<script src="/assets/js/jquery-3.1.1.min.js"></script>
-<script src="/assets/bootstrap-4/js/bootstrap.min.js"></script>
-<script src="/assets/js/input.js"></script>
-</html>
-
-{{end}}
