+++ assets/template/import-scenario-pdf.html
@@ -50,12 +50,6 @@
 <script src="/assets/bootstrap-5.0.2/js/bootstrap.bundle.min.js"></script>
 
 <script src="/assets/js/upload-scenario-pdf.js"></script>
-<!--
-<script src="/assets/js/dropzone.js"></script>
-<script src="/assets/js/input.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
-<script src="/assets/js/export.js"></script>
--->
 </html>
 
 {{end}}
