+++ assets/template/importjson.html
@@ -37,7 +37,6 @@
 <script src="/assets/js/jquery-3.1.1.min.js"></script>
 <script src="/assets/bootstrap-4/js/bootstrap.min.js"></script>
 <script src="/assets/js/input.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
 <script src="/assets/js/dropzone.js"></script>
 <script src="/assets/js/export.js"></script>
 </html>
