+++ assets/template/index.html
@@ -30,7 +30,6 @@
 <script src="/assets/js/input.js"></script>
 <script src="/assets/js/Sortable.js"></script>
 <script src="/assets/js/export.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
 <script src="/assets/js/scroll.js"></script>
 <script src="/assets/js/dropzone.js"></script>
 <script>
