+++ assets/template/reportjson.html
@@ -80,7 +80,6 @@
 <script src="/assets/js/jquery-3.1.1.min.js"></script>
 <script src="/assets/bootstrap-4/js/bootstrap.min.js"></script>
 <script src="/assets/js/input.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
 <script src="/assets/js/dropzone.js"></script>
 </html>
