+++ assets/template/reviewstatus.html
@@ -188,7 +188,6 @@
 <script src="/assets/js/jquery-3.1.1.min.js"></script>
 <script src="/assets/bootstrap-5.0.2/js/bootstrap.min.js"></script>
 <script src="/assets/js/input.js"></script>
-<script src="/assets/js/opio_v02.js"></script>
 <script src="/assets/js/dropzone.js"></script>
 <script type="text/javascript">
     selectReviewItem('{{.CurrentReview.ID.Hex}}')
